refactor: exception 폴더 생성해서 그룹화
