+++ src/main/java/roomescape/exception/exception/BaseCustomException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exception;
 
 public class BaseCustomException extends RuntimeException implements CustomException {
     private final ErrorCode errorCode;
