+++ src/main/java/roomescape/exception/exception/CustomException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exception;
 
 public interface CustomException {
     ErrorCode getErrorCode();
