+++ src/main/java/roomescape/exception/exception/DataReferencedException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exception;
 
 public class DataReferencedException extends BaseCustomException {
     public DataReferencedException(ErrorCode errorCode) {
