+++ src/main/java/roomescape/exception/exception/DuplicatedResourceException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exception;
 
 public class DuplicatedResourceException extends BaseCustomException {
     public DuplicatedResourceException(ErrorCode errorCode) {
