+++ src/main/java/roomescape/exception/exception/NotFoundResourceException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.exception.exception;
 
 public class NotFoundResourceException extends BaseCustomException {
     public NotFoundResourceException(ErrorCode errorCode) {
