+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -7,16 +7,15 @@ import roomescape.domain.reservationTime.ReservationTime;
 import roomescape.domain.reservationTime.ReservationTimeCondition;
 import roomescape.domain.reservationTime.ReservationTimeWithAvailable;
 import roomescape.dto.reservationTime.AddReservationTimeRequest;
-import roomescape.exception.DataReferencedException;
-import roomescape.exception.DuplicatedResourceException;
-import roomescape.exception.ErrorCode;
+import roomescape.exception.exception.DataReferencedException;
+import roomescape.exception.exception.DuplicatedResourceException;
 import roomescape.repository.reservation.ReservationRepository;
 import roomescape.repository.reservationTime.ReservationTimeRepository;
 
 import java.util.List;
 
-import static roomescape.exception.ErrorCode.*;
-import static roomescape.exception.ErrorCode.DUPLICATED_RESERVATION_TIME;
+import static roomescape.exception.exception.ErrorCode.*;
+import static roomescape.exception.exception.ErrorCode.DUPLICATED_RESERVATION_TIME;
 
 @Service
 public class ReservationTimeService {
