+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -8,9 +8,8 @@ import roomescape.domain.theme.Theme;
 import roomescape.domain.theme.ThemeWithCount;
 import roomescape.dto.theme.AddThemeRequest;
 import roomescape.dto.theme.PopularConditionRequest;
-import roomescape.exception.DataReferencedException;
-import roomescape.exception.DuplicatedResourceException;
-import roomescape.exception.ErrorCode;
+import roomescape.exception.exception.DataReferencedException;
+import roomescape.exception.exception.DuplicatedResourceException;
 import roomescape.repository.reservation.ReservationRepository;
 import roomescape.repository.theme.ThemeRepository;
 
@@ -20,7 +19,7 @@ import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static roomescape.exception.ErrorCode.*;
+import static roomescape.exception.exception.ErrorCode.*;
 
 public class ThemeServiceTest {
     private ReservationRepository createReservationRepository(boolean isExistTheme) {
